Made a component for the email input fields to avoid doubled code
